refactor(#133 WI-9): split sheet UI into <300-line files
Keeps each file focused + under the ~300-line guideline: the query bar moves to
InBookSearchField.kt, the Indexing/NoResults body states to InBookSearchStates.kt;
InBookSearchSheet.kt (271) keeps the sheet chrome + body dispatch + results list,
InBookSearchRows.kt (275) keeps the summary/group-header/hit/recent rows. Unused
imports pruned. Behavior-preserving; the androidTest set compiles + JVM suite green.

Refs #133
