+++ android/app/src/main/kotlin/com/vreader/app/search/InBookSearchSheet.kt
@@ -15,8 +15,6 @@
 package com.vreader.app.search
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.focusable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -25,17 +23,10 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.BasicTextField
-import androidx.compose.foundation.text.KeyboardOptions
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Close
-import androidx.compose.material.icons.filled.Search
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
@@ -43,19 +34,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.snapshotFlow
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.graphics.SolidColor
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.vreader.app.reader.nav.JumpResult
@@ -133,7 +115,7 @@ fun InBookSearchSheetContent(
             .background(theme.background)
             .testTag("inbook-search-sheet-content"),
     ) {
-        SearchField(
+        InBookSearchField(
             theme = theme,
             bookTitle = bookTitle,
             query = query,
@@ -170,97 +152,6 @@ fun InBookSearchSheetContent(
             is InBookSearchContent.Error -> Unit
             InBookSearchContent.Unsupported -> Unit
         }
-    }
-}
-
-/**
- * The design's search bar: a rounded pill (search glyph · autofocus text field · inline clear) + the
- * trailing "Cancel". The field auto-focuses on first composition (the design's `inputRef.focus()`), so the
- * keyboard is up the moment the sheet opens. Clearing (the inline ✕) resets the query via [onQueryChange].
- */
-@Composable
-private fun SearchField(
-    theme: ReaderTheme,
-    bookTitle: String,
-    query: String,
-    onQueryChange: (String) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val ink = theme.ink
-    val sub = ink.copy(alpha = 0.55f)
-    val fieldFill = ink.copy(alpha = if (theme.isDark) 0.06f else 0.05f)
-    val focusRequester = remember { FocusRequester() }
-    val hasQuery = query.isNotEmpty()
-    val shape = RoundedCornerShape(12.dp)
-
-    // Autofocus on first composition — the design's `setTimeout(() => inputRef.focus())`.
-    LaunchedEffect(Unit) {
-        runCatching { focusRequester.requestFocus() }
-    }
-
-    // The design's single rounded pill: search glyph · input · clear · Cancel, all inside one container
-    // (vreader-search.jsx lines 55-79) — no invented accent border.
-    Row(
-        Modifier
-            .fillMaxWidth()
-            .padding(start = 16.dp, end = 16.dp, top = 12.dp, bottom = 8.dp)
-            .clip(shape)
-            .background(fieldFill)
-            .padding(horizontal = 14.dp, vertical = 10.dp),
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(Icons.Filled.Search, contentDescription = null, tint = sub, modifier = Modifier.size(18.dp))
-        BasicTextField(
-            value = query,
-            onValueChange = onQueryChange,
-            modifier = Modifier
-                .weight(1f)
-                .focusRequester(focusRequester)
-                .focusable()
-                .testTag("inbook-search-field")
-                .semantics { contentDescription = "Search this book" },
-            singleLine = true,
-            textStyle = TextStyle(color = ink, fontFamily = VReaderFonts.Sans, fontSize = 15.sp),
-            cursorBrush = SolidColor(theme.accent),
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-            decorationBox = { inner ->
-                if (!hasQuery) {
-                    Text(
-                        "Search $bookTitle",
-                        color = sub,
-                        fontFamily = VReaderFonts.Sans,
-                        fontSize = 15.sp,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                }
-                inner()
-            },
-        )
-        if (hasQuery) {
-            Icon(
-                Icons.Filled.Close,
-                contentDescription = "Clear search",
-                tint = sub,
-                modifier = Modifier
-                    .size(16.dp)
-                    .clickable { onQueryChange("") }
-                    .testTag("inbook-search-clear"),
-            )
-        }
-        Text(
-            "Cancel",
-            modifier = Modifier
-                .clickable(onClick = onDismiss)
-                .padding(start = 6.dp)
-                .testTag("inbook-search-cancel")
-                .semantics { contentDescription = "Cancel search" },
-            color = theme.accent,
-            fontFamily = VReaderFonts.Sans,
-            fontSize = 14.sp,
-            fontWeight = FontWeight.Medium,
-        )
     }
 }
